+++ frontend/app/backend-strings.generated.js
@@ -217,6 +217,7 @@ export const backendMappingKeys = [
   'backend_mappings.trade_location.optimism',
   'backend_mappings.trade_location.polkadot',
   'backend_mappings.trade_location.poloniex',
+  'backend_mappings.trade_location.robinhood',
   'backend_mappings.trade_location.polygon_pos',
   'backend_mappings.trade_location.realestate',
   'backend_mappings.trade_location.scroll',