+++ rotkehlchen/chain/balances.py
@@ -47,6 +47,7 @@ class BlockchainBalances:
     binance_sc: defaultdict[ChecksumEvmAddress, BalanceSheet] = field(init=False)
     monad: defaultdict[ChecksumEvmAddress, BalanceSheet] = field(init=False)
     sonic: defaultdict[ChecksumEvmAddress, BalanceSheet] = field(init=False)
+    robinhood: defaultdict[ChecksumEvmAddress, BalanceSheet] = field(init=False)
     eth2: defaultdict[Eth2PubKey, BalanceSheet] = field(init=False)
     btc: dict[BTCAddress, Balance] = field(init=False)
     bch: dict[BTCAddress, Balance] = field(init=False)