+++ rotkehlchen/chain/evm/decoding/weth/constants.py
@@ -1,6 +1,13 @@
 from typing import Final
 
-from rotkehlchen.constants.assets import A_WETH, A_WETH_ARB, A_WETH_BASE, A_WETH_OPT, A_WETH_SCROLL
+from rotkehlchen.constants.assets import (
+    A_WETH,
+    A_WETH_ARB,
+    A_WETH_BASE,
+    A_WETH_OPT,
+    A_WETH_ROBINHOOD,
+    A_WETH_SCROLL,
+)
 from rotkehlchen.types import ChainID
 
 CPT_WETH: Final = 'weth'
@@ -12,11 +19,12 @@ CHAINS_WITHOUT_NATIVE_ETH: Final = {
     ChainID.MONAD,
     ChainID.SONIC,
 }
-CHAINS_WITH_SPECIAL_WETH: Final = {ChainID.SCROLL, ChainID.ARBITRUM_ONE, ChainID.BASE}
+CHAINS_WITH_SPECIAL_WETH: Final = {ChainID.SCROLL, ChainID.ARBITRUM_ONE, ChainID.BASE, ChainID.ROBINHOOD}  # noqa: E501
 CHAIN_ID_TO_WETH_MAPPING: Final = {
     ChainID.ETHEREUM: A_WETH,
     ChainID.ARBITRUM_ONE: A_WETH_ARB,
     ChainID.OPTIMISM: A_WETH_OPT,
     ChainID.SCROLL: A_WETH_SCROLL,
     ChainID.BASE: A_WETH_BASE,
+    ChainID.ROBINHOOD: A_WETH_ROBINHOOD,
 }