+++ rotkehlchen/constants/location_details.py
@@ -81,6 +81,7 @@ LOCATION_DETAILS: dict = {
     Location.BINANCE_SC: {'image': 'binance_sc.svg'},
     Location.MONAD: {'image': 'monad.svg'},
     Location.SONIC: {'image': 'sonic.svg'},
+    Location.ROBINHOOD: {'image': 'robinhood.svg'},
     Location.AVALANCHE: {'image': 'avalanche.svg'},
     Location.WOO: {'image': 'woo.svg'},
     Location.BYBIT: {'image': 'bybit.svg'},