+++ rotkehlchen/db/schema.py
@@ -122,6 +122,8 @@ INSERT OR IGNORE INTO location(location, seq) VALUES ('y', 57);
 INSERT OR IGNORE INTO location(location, seq) VALUES ('z', 58);
 /* Sonic */
 INSERT OR IGNORE INTO location(location, seq) VALUES ('~', 62);
+/* Robinhood */
+INSERT OR IGNORE INTO location(location, seq) VALUES ('\x7f', 63);
 /* Gate */
 INSERT OR IGNORE INTO location(location, seq) VALUES ('{', 59);
 /* Bit2me */