+++ rotkehlchen/tests/utils/ethereum.py
@@ -27,6 +27,8 @@ from rotkehlchen.chain.optimism.decoding.decoder import OptimismTransactionDecod
 from rotkehlchen.chain.optimism.transactions import OptimismTransactions
 from rotkehlchen.chain.polygon_pos.decoding.decoder import PolygonPOSTransactionDecoder
 from rotkehlchen.chain.polygon_pos.transactions import PolygonPOSTransactions
+from rotkehlchen.chain.robinhood.decoding.decoder import RobinhoodTransactionDecoder
+from rotkehlchen.chain.robinhood.transactions import RobinhoodTransactions
 from rotkehlchen.chain.scroll.decoding.decoder import ScrollTransactionDecoder
 from rotkehlchen.chain.scroll.transactions import ScrollTransactions
 from rotkehlchen.chain.sonic.decoding.decoder import SonicTransactionDecoder
@@ -459,6 +461,7 @@ def get_decoded_events_of_transaction(
         ChainID.BINANCE_SC: (BinanceSCTransactions, BinanceSCTransactionDecoder),
         ChainID.MONAD: (MonadTransactions, MonadTransactionDecoder),
         ChainID.SONIC: (SonicTransactions, SonicTransactionDecoder),
+        ChainID.ROBINHOOD: (RobinhoodTransactions, RobinhoodTransactionDecoder),
         ChainID.HYPERLIQUID: (HyperliquidTransactions, HyperliquidTransactionDecoder),
     }
     mappings_result = chain_mappings.get(evm_inquirer.chain_id)